requirements.txt
PySide6
sqlmodel
python-dateutil
APScheduler
win10toast_click
pydantic-settings
pywin32